requirements.txt
flask
gunicorn
flask-cors
python-dotenv
ntplib
tzdata
requests
pymongo